data_collector/requirements.txt
flask
PyYaml
apscheduler
celery
requests
redis